Decode char strings the same as a stream of single chars
The 'if else' checks for special characters are refactored into a separate
function so it can be applied to each character when looping over them
in the 'fall through' case. Streaming a character at a time is now the
same as concatenating the characters before writing to the input stream.